+++ src/lang-xpath/main/uk/co/reecedunn/intellij/plugin/xpath/ast/xpath/XPathLocalUnionType.kt
@@ -17,7 +17,6 @@ package uk.co.reecedunn.intellij.plugin.xpath.ast.xpath
 
 import com.intellij.psi.PsiElement
 import uk.co.reecedunn.intellij.plugin.xdm.types.XdmItemType
-import uk.co.reecedunn.intellij.plugin.xdm.types.XsQNameValue
 
 /**
  * An XPath 4.0 ED and XQuery 4.0 ED `LocalUnionType` node in the XPath/XQuery AST.
@@ -25,5 +24,5 @@ import uk.co.reecedunn.intellij.plugin.xdm.types.XsQNameValue
  * Saxon 9.8 supports an earlier version of this.
  */
 interface XPathLocalUnionType : PsiElement, XdmItemType {
-    val memberTypes: Sequence<XsQNameValue>
+    val memberTypes: Sequence<XdmItemType>
 }