+++ src/lang-xquery/test/uk/co/reecedunn/intellij/plugin/xquery/tests/parser/XQueryParserTest.kt
@@ -9263,22 +9263,6 @@ private class XQueryParserTest : ParserTestCase() {
         @DisplayName("XQuery 4.0 ED EBNF (233) LocalUnionType")
         internal inner class LocalUnionType {
             @Test
-            @DisplayName("NCName")
-            fun ncname() {
-                val expected = loadResource("tests/parser/xquery-4.0/LocalUnionType_NCName.txt")
-                val actual = parseResource("tests/parser/xquery-4.0/LocalUnionType_NCName.xq")
-                assertThat(actual.toPsiTreeString(), `is`(expected))
-            }
-
-            @Test
-            @DisplayName("NCName; compact whitespace")
-            fun ncname_CompactWhitespace() {
-                val expected = loadResource("tests/parser/xquery-4.0/LocalUnionType_NCName_CompactWhitespace.txt")
-                val actual = parseResource("tests/parser/xquery-4.0/LocalUnionType_NCName_CompactWhitespace.xq")
-                assertThat(actual.toPsiTreeString(), `is`(expected))
-            }
-
-            @Test
             @DisplayName("qname")
             fun qname() {
                 val expected = loadResource("tests/parser/xquery-4.0/LocalUnionType.txt")
@@ -9291,22 +9275,6 @@ private class XQueryParserTest : ParserTestCase() {
             fun qname_CompactWhitespace() {
                 val expected = loadResource("tests/parser/xquery-4.0/LocalUnionType_CompactWhitespace.txt")
                 val actual = parseResource("tests/parser/xquery-4.0/LocalUnionType_CompactWhitespace.xq")
-                assertThat(actual.toPsiTreeString(), `is`(expected))
-            }
-
-            @Test
-            @DisplayName("URIQualifiedName")
-            fun uriQualifiedName() {
-                val expected = loadResource("tests/parser/xquery-4.0/LocalUnionType_URIQualifiedName.txt")
-                val actual = parseResource("tests/parser/xquery-4.0/LocalUnionType_URIQualifiedName.xq")
-                assertThat(actual.toPsiTreeString(), `is`(expected))
-            }
-
-            @Test
-            @DisplayName("URIQualifiedName; compact whitespace")
-            fun uriQualifiedName_CompactWhitespace() {
-                val expected = loadResource("tests/parser/xquery-4.0/LocalUnionType_URIQualifiedName_CompactWhitespace.txt")
-                val actual = parseResource("tests/parser/xquery-4.0/LocalUnionType_URIQualifiedName_CompactWhitespace.xq")
                 assertThat(actual.toPsiTreeString(), `is`(expected))
             }